Added post_show_page.rb file to the feature folder
